fix(parser): let the arity branch settle the slot, and file the DSL as a fixture
Review found the shape checks in `TensorOptionalSlotPattern` redundant. They
were added before the layout branch carried its lookahead, when a branch had to
refuse a slot it had already been handed. The lookahead now settles which branch
runs, so the role reaching this pattern is already decided and asking again only
obscured where the decision is made. Removing them leaves the two
`layout_or_storage` branches with no caller, so they go too; `_slot_kind`
remains as the lookahead's classifier.

The runtime-trip program moves out of `smoke_analyze.py` into
`tests/fixtures/diagnostics/dynamic_trip_count.py`, beside the tuple-projection
diagnostic it is a sibling of, and reaches the test through a conftest accessor
like every other fixture.

source 927 passed / 4 skipped, installed 168 passed.
